requirements.txt
PySimpleGUI~=4.57.0
PySimpleGUIQt~=0.35.0
psutil~=5.9.0
Pillow~=9.0.1
cx_freeze